Restore some changes from older commits.

+++ ctf/g_ai.c
@@ -55,7 +55,7 @@ void AI_SetSightClient (void)
 	if (level.sight_client == NULL)
 		start = 1;
 	else
-		start = level.sight_client - g_edicts;
+		start = (int)(level.sight_client - g_edicts);
 
 	check = start;
 	while (1)
