+++ ctf/g_utils.c
@@ -405,7 +405,7 @@ void G_InitEdict (edict_t *e)
 	e->inuse = true;
 	e->classname = "noclass";
 	e->gravity = 1.0;
-	e->s.number = e - g_edicts;
+	e->s.number = (int)(e - g_edicts);
 }
 
 /*
